Add files via upload
Q1. What is an activation function in the context of artificial neural networks?

Q2. What are some common types of activation functions used in neural networks?

Q3. How do activation functions affect the training process and performance of a neural network?

Q4. How does the sigmoid activation function work? What are its advantages and disadvantages?

Q5.What is the rectified linear unit (ReLU) activation function? How does it differ from the sigmoid function?

Q6. What are the benefits of using the ReLU activation function over the sigmoid function?

Q7. Explain the concept of "leaky ReLU" and how it addresses the vanishing gradient problem.

Q8. What is the purpose of the softmax activation function? When is it commonly used?

Q9. What is the hyperbolic tangent (tanh) activation function? How does it compare to the sigmoid function?