Add files via upload
Q1. What is the purpose of forward propagation in a neural network?

Q2. How is forward propagation implemented mathematically in a single-layer feedforward neural network?

Q3. How are activation functions used during forward propagation?

Q4. What is the role of weights and biases in forward propagation?

Q5. What is the purpose of applying a softmax function in the output layer during forward propagation?

Q6. What is the purpose of backward propagation in a neural network?

Q7. How is backward propagation mathematically calculated in a single-layer feedforward neural network?

Q8. Can you explain the concept of the chain rule and its application in backward propagation?

Q9. What are some common challenges or issues that can occur during backward propagation, and how
can they be addressed?